web_app/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
django-filter==23.3
python-jose==3.3.0




